kernel - merge m_pulldown() bug fix
In m_pulldown(), before trying to prepend bytes to the subsequent mbuf,
ensure that the subsequent mbuf contains the remainder of the bytes
the caller sought. If this is not the case, fall through to the code
which gathers the bytes in a new mbuf.

This fixes a bug where m_pulldown() could fail to gather all the desired
bytes into consecutive memory.

Patch-by:	sephe
FreeBSD-PR:	238787
Reported by:	A reddit user
Discussed with: emaste
Obtained from:	NetBSD
Obtained from:	FreeBSD