route: Don't install a cloned route when interface is down
When an interface is taken down, all subnet routes are removed
and thus all cloned routes are flushed.

It's possible, although unlikely, for userland to add a subnet
route when the interface is down.

In this state, EHOSTDOWN would be returned regardless, but a
cloned route would be created locked to the interface.
Instead, check if interface is down before creating the cloned
route so it's not made needlessly.