inet6: Restore ND6_LLINFO_WAITDELETE
This predates DragonFlyBSD by quite a few years, but
ND6_LLINFO_WAITDELETE was removed in 2000 from all the BSD's.
This change brings it back so that if someone *really* wants
to resolve the unresolvable we don't constant delete
and add cloned routes which is just a waste of resources.

It also helps to return EHOSTDOWN more consistently to senders.
If we don't resolve the address during the normal wait cyckle OR
attempt to resolve it again while in the ND6_LLINFO_WAITDELETE
state, then then cloned route is then freed.