login.conf: Don't dump /COPYRIGHT on users' screens when logging in.
This reverts e72f0762b947abfb33852312cbec661ce8761559.

The /etc/COPYRIGHT file is meant to be created by the local admin
and to be non-existent by default. It has nothing to do with our
/COPYRIGHT, except the name.

It looks like FreeBSD went through the same cycle. See their
revisions 70122 and 70189.

Also add a comment that explains what's intended.

Discussed-with: aly