dhcpcd: Remove dependency on netifs
dhcpcd currently depends on being able to bind to the
unspecified address on the BOOTPC port and DHCPv6-client port
as well as listen to ICMP6.
This means that some routes need to be established first and
the rc.d infrastructure relies on netif being started before that.

We need to document not to configure an interface to start DHCP,
but rather just to start dhcpcd and if any extra configuration is
needed to limit the scope of interfaces it should be done in
dhcpcd.conf instead. However, the default should be fine.