if_clone: Extend if_clone_create() by an extra parameter
Currently, if_clone_create() accepts one caddr_t parameter that is
passed with the ifreq->ifr_data for the SIOCIFCREATE2 ioctl.

Extend this function by another caddr_t parameter so that callers could
pass extra data to the ifc->ifc_create() handler and won't conflict with
the SIOCIFCREATE2 ioctl calls.  This extension will be used by tun(4)
and tap(4) in a following commit.

Suggested-by: dillon
Reviewed-by: dillon