Bring in FreeBSD's wsp(4), for Wellspring touchpads on Apple laptops.
This commit also removes ums(4) from the default kernel config file,
because it would otherwise claim this device. It is safe to remove,
and in fact FreeBSD also doesn't have it in their GENERIC, because
devd(8) will properly take care of loading the appropriate modules
for both ums(4) and wsp(4) and starting a moused(8).

Taken-from: FreeBSD
Tested-by:  Patrick McDonough <~patrick/dfly@awk.is>