kernel/iscsi: Remove iscsi's local definition of boolean_t.
It was locally defining boolean_t as int (4 bytes) for userland, but
boolean_t is _Bool (1 byte) in the kernel since quite some time.

This caused isc_opt_t's size being different in userland vs. kernel,
and ultimately caused the ioctl number of ISCSISETOPT, which includes
sizeof(isc_opt_t), to be different in userland vs. the kernel,
resulting in at first inexplicable ENOIOCTL issues when using
iscontrol(8).

Change all 'boolean_t' usage to 'bool' which is available for both
userland and kernel, defined as _Bool.

Thanks-to: Georg "megaT" Bege <georg@bege.email> for spotting the ioctl
           number difference & testing