vendor/ncurses: upgrade from 6.0-20160305 to 6.2-20200212
The TERMINAL structure in <term.h> is now opaque. Doing that allowed
making the structure larger, to hold the extended numeric data.
When configured for wide-characters (i.e. ncursesw), the TERMINAL
structure is extended and is named TERMTYPE2. When configured for
8-bit (narrow) characters, the TERMTYPE2 structure is not used.

For detailed description of all other changes, enhancements, bugfixes
and performance improvements, see:

https://invisible-island.net/ncurses/announce-6.1.html
https://invisible-island.net/ncurses/announce-6.2.html