Stop mixing the userland specific <bsd.sys.mk> into the module build.
I don't know why it was there initially because we inherited it this way.

The only changes on the compile line are eliminating a -std=gcc99
(our userland -std setting) which was overridden by the kernel's
own -std=c99 anyway, and a second copy of our kernel warning flags.
By extension, it also makes the build log shorter by some percent.