Clean up the POSIX namespace of <iconv.h>, <dirent.h>, and <sys/stat.h>.
Some small residue in <sys/stat.h> remains for older specs, due to the
kind of bumpy history of that header between issues 6 and 7. Namely,
struct timespec should not be defined in these cases. But this is good
enough for now.

Thanks to zrj for testing with a full dports bulk.