Import stdbuf(1) together with libstdbuf(3) from FreeBSD
The stdbuf(1) is a utility to change the initial buffering of stdin,
stdout and stderr streams for a given command.  It achieves this
functionality by preloading the libstdbuf(3) library and configuring the
needed environment variables.  Under the hood, libstdbuf(3) library
calls setvbuf(3) to do the actual work.

Changes from FreeBSD:
* Removed 32bit i386 code from stdbuf(1)
* Fix typos and update HISTORY section in libstdbuf.3 man page
* Update the example to actually work on DragonFly