Import getaddrinfo(1) from FreeBSD
The getaddrinfo(1) utility resolves host and service names to socket
addresses with getaddrinfo(3) and prints them in a user-friendly format.

Obtained from FreeBSD, which obtained it from NetBSD.

The examples in the man page has been adjusted to use
'www.dragonflybsd.org' instead of NetBSD ones :D

Thank swildner for bringing in sockaddr_snprintf(3) to libutil from
NetBSD, which is required by this utility.