talkd: Find users in more than one talk request.
talkd was written to use `utmpentry` from `who`, which
caches the results of read the `utmpx` file.  However,
talkd is usually invoked from inetd `wait` mode; it's
possible that a user might login after `talkd` starts;
with the cached utmpx data, one can't `talk` to that
user.

Further, consumption of the utmp data nulls it out for
subsequent requests.  The result is that the first talk
succeeds, but subsequent requests fail.

The fix is to avoid using the `utmpentry` machinery,
and just read `utmp` directly every time we need to.

Signed-off-by: Dan Cross <cross@gajendra.net>