ching(6): Support to format with mandoc(1)
We don't ship nroff(1) anymore, so adapt ching(6) to support to use
mandoc(1) to format its output.  Mandoc(1) only implements part of
roff(7) language, but it's enough for ching(6), although minor
adjustments to the mandoc(1) output are needed to make it similar to
the nroff(1) output.

Prefer nroff(1) if it's available (provided by groff package).