games: Bring in monop(6) from NetBSD
monop is reminiscent of the Parker Brother's game Monopoly, and monitors
a game between 2 to 9 users.  It is assumed that the rules of Monopoly
are known.  The game follows the standard rules, with the exception that,
if a property goes up for auction and there are only two solvent players,
no auction is held and the property remains unowned.