cgram(6): Don't move over the last char and fix an assertion
The maximum value of cursor_x was the text length and thus also
*equaled* to extent_x, which broke the assertion in substitute()
and caused crashes.

In addition, the cursor could move just beyond the last character
on a line, which didn't make much sense.

So only allow the cursor move to the last char but not over it.
This also properly fix the above assertion issue.