od(1): Fix long ints (L) format output on x86_64 platform
Long ints are of size 8 bytes on x86_64 platform, and the handle of
8-byte integers was missing in od(1)/hexdump(1).  So using '-t [doux]L'
(long ints) options in od(1) resulted in error:

od: d: bad byte count

Update the parse code to handle 8-byte long integers to fix it, and
update odformatint() to correctly calculate the digits.
In addition, fix the type error of 's8' (int8_t -> int64_t) in print(),
otherwise, long ints are truncated and get wrong.

Reported-by: swildner