kernel/mptable: Fix wrong start of mptable search area for non-EBDA case.
If no Extended BIOS Data Area (EBDA) exists, the kernel will fall back
to searching the MP table in the last kilobyte of the base memory. The
amount of base memory is in kilobytes but the calculation was treating
it like bytes. Adjust the calculation.

Nothing except bhyve seems to have triggered this path so far. Usually
there is an EBDA.

Thanks to everyone involved.

Reported-by: Pierre-Alain TORET <pierre-alain.toret@protonmail.com>
Fix-by:      Peter Grehan <grehan@freebsd.org> (changed a little by me)
FreeBSD-bug: https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=255191