devfs: Change typedef 'cdevpriv_dtr_t' to 'd_priv_dtor_t'
The original 'cdevpriv_dtr_t' typedef was not able to be used in a
function prototype like the other d_xxx_t typedef's, because it declared
a function *pointer* rather than a function.  This makes it useless
outside the cdevpriv implementation.

Change the typedef to declare a function and rename it to
'd_priv_dtor_t', which is more consistent with other dev_ops methods,
although it's not a direct member in the dev_ops struct.

Obtained-from: FreeBSD (revision 291653)
URL: https://reviews.freebsd.org/D4340