nvmm: Port to DragonFly #12: FPU save & restore
Note that the host FPU state is indeterminant and depends on whether
the user program used the FPU or not, so there might not be any state to
save.  npxpush() and npxpop() can handle this.  Accordingly, need to use
'mcontext_t' to store host FPU state.

At first I used fpu_area_save() and fpu_area_restore() to deal with the
host FPU state, but it caused a hard fault loop when trying to boot an
OS in QEMU, because it failed to handle an uninitialized FPU.  Thanks
to Matt Dillon for tracking it down and fixing it.

Credit to FreeBSD vmm code: save_guest_fpustate(), restore_guest_fpustate()