test/nvmm: Add a Makefile and a test script for 'calc-vm'
We write the built binary in /tmp instead of current directory.  This is
useful for building with a NFS-exported readonly mount of the source
tree (e.g., used by dillon).

Also add a test script that runs the 'calc-vm' test program in a loop.
It helped reveal the VMCS remote clear bug.