getconf(1): Add _POSIX_V7_* names, as aliases for POSIX_V7_*.
The standard is indeed ambiguous about whether the confstr(3) related
POSIX_Vx_* names shall have a leading underscore or not. I think the
intention is to follow the specified rule:

"The names of the symbolic constants used as the name argument of the
 confstr() function in the System Interfaces volume of POSIX.1-2017,
 without the _CS_ prefix."

But in its description of the -v option it uses leading underscores.