kernel: Remove ndis(4) and associated tools and stuff.
ndis(4) was a wrapper to allow running binary Windows network
drivers that conformed to the Network Driver Interface Specification,
i.e. NDIS.

It only ever supported drivers from the days of Windows XP and
Windows Server 2003 (i.e. NDIS 5.1). And even if one was actually
able to extract the .sys and .inf files from the driver package and
successfully convert them to a building module, which both were
adventures in itself, it could be any result when trying to run it,
depending on the card and driver, from resonably working to not
working, even crashing. But it did work for some cards, so it had
some limited merit in its time.

NetBSD removed it in 2018, FreeBSD in January 2021, so let's follow
suit now.