nvmm: Implement os_ipi_broadcast() for _change_cpu()
os_ipi_broadcast() is used in {svm,vmx}_init() and {svm,vmx}_fini() to
enable/disable SVM/VMX on all CPUs.  The previous implementation used an
interlock and made significant changes to the code.  Reimplement it in
another way (loop on each CPU and execute the task), so it can be easily
enclosed in the os_ipi_broadcast() function.

The new implementation referred to our pc64/vmm/vmx.c code.