syslogd(8): Fix klog may be broken into multiple incomplete lines
Fix readklog() to not emit partial lines by changing it to use a static
buffer for reading log lines from /dev/klog.  Thus we can make it
only emit the line for printing if there is a complete line.

This fixes the issue that a kernel log line may be broken into multiple
incomplete lines that are written into log file by syslogd(8),
especially when the system is under load.

Patch-by: Matt Dillon