<errno.h>: Make ENOTRECOVERABLE and EOWNERDEAD visible unconditionally.
The problem here is that GCC will expect these errnos to be
visible unconditionally for its C++ headers - and errors in
<bits/error_constants> if they aren't - as soon as we set the
HAVE_E* for them in GCC's config.h.

On the other hand we have software such as www/qt5-webengine
that sets _POSIX_C_SOURCE and/or _XOPEN_SOURCE to older standards
that didn't specify those two errnos.

If anything, it is rather a "fault" of GCC's C++ header handling
than one of qt5-webengine.

Reported-by: tuxillo