installer(8): Move the contents of sysids.h into <sys/diskmbr.h>.
It is needed by both fdisk(8) and the installer.

<sys/diskmbr.h> seems reasonable because it already has defines for
various partition types. Use them in the struct. Adjust the naming of
the structure and the array to match the rest of <sys/diskmbr.h>.

Based on a patch by Alexander Kuleshov <kuleshovmail@gmail.com>.

Dragonfly-bug: <https://bugs.dragonflybsd.org/issues/2875>