getopt.3: Clarify option string syntax
The recently added 'option::' syntax requires the following argument
(if present) *must* be in the same word as the option name itself
(e.g., "-x<arg>").  This is different than the normal 'option:' syntax,
for which the argument may have a leading white space (e.g., "-x <arg>").
Update the man page to explicitly describe the syntax.

Update to use a list to describe the 3 types of option string syntaxes
with examples.