tools/pw-update.sh: Allow to create user 'toor' of duplicate UID
The 'toor' user has duplicate UID of 0 as 'root', and it also has an
empty shell field.  This caused problem for the 'pw-update.sh' to
handle.  If the user/admin deliberately removed the 'toor' user, then
this script was unable to add it back, which breaks the 'preinstall'
target and thus 'installworld' failed.

Use the '-o' option for 'pw useradd' to allow to create users of
duplicate UIDs.  In addition, quote the shell field to handle the empty
case.

Bug: https://bugs.dragonflybsd.org/issues/3332