pw(8): Allow to change GID of non-root super users (UID=0)
pw(8) was silently refusing to change the GID/group of super users (i.e.,
UID=0), and thus couldn't change the group of 'toor' user.

Relax the policy a bit and only preventing from changing the GID/group
of 'root' user, and raise an error in that case.  The 'root' is regarded
as the special super user and already has several special checks in the
code.  So this change actually make the behavior more consistent.

This commit fixes the special case of 'tools/pw-update.sh' adding the
'toor' user.  It first adds the user with group of 'nogroup' and later
change to its real group.

Bug: https://bugs.dragonflybsd.org/issues/3332