top(1): Fix count of process states
Previously, the 'active' (SACTIVE) process state includes the 'running'
(LSRUN) state.  This caused the total process count was less than the
sum of all process states.

A process consists of one or more LWPs.  The process state (kp_stat)
is more of a meta-state for the process, which can be:
- SIDL: starting
- SACTIVE: active
- SSTOP: stopped
- SZOMB: zombie
- SCORE: about to die (performing a core dump)

For an active process (SACTIVE), the LWP state (kp_lwp.kl_stat)
represents the actual state, which can be:
- LSRUN: running
- LSSTOP: stopped
- LSSLEEP: sleeping

Explicitly check for all the above states and count them, and thus
obtain the correct count of each process state.