kernel/linker: Resolve module's relative path to full path
Before this change, the relative path was stored and shown in
'kldstat -v' output when a relative path was used to load it.
That behavior wasn't actually expected since we really want the full
paths in 'kldstat -v' output to avoid any confusion.  So resolve the
relative paths to absolute full paths.